refactor: rename comply-with-me → CompliGator (v0.2.0-beta) (#47)
- comply_with_me.py → compligator.py
- cwm/ package → compligator/
- All from cwm. imports → from compligator.
- .cwm-venv/ → .compligator-venv/ (venv recreated fresh on first run)
- .cwm-state.json → .compligator-state.json
- pyproject.toml: name=compligator, version=0.2.0, scripts entry=compligator
- .github/workflows/lint.yml: ruff check compligator/
- README: full rewrite for CompliGator identity + new frameworks
- source-content placeholder READMEs: updated references
