+++ .gitignore
@@ -54,6 +54,7 @@ pnpm-debug.log*
 # Python tooling
 .venv/
 .cwm-venv/
+.compligator-venv/
 venv/
 .pytest_cache/
 .mypy_cache/
