+++ pyproject.toml
@@ -3,9 +3,9 @@ requires = ["setuptools>=68", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "comply-with-me"
-version = "0.3.0"
-description = "CLI tool for downloading and syncing compliance framework documentation"
+name = "compligator"
+version = "0.2.0"
+description = "Compliance document aggregator — download, sync, and normalize framework docs"
 readme = "README.md"
 requires-python = ">=3.9"
 dependencies = [
@@ -16,11 +16,11 @@ dependencies = [
 ]
 
 [project.scripts]
-cwm = "cwm.cli:main"
+compligator = "compligator.cli:main"
 
 [tool.setuptools.packages.find]
 where = ["."]
-include = ["cwm*"]
+include = ["compligator*"]
 
 [tool.ruff]
 target-version = "py39"
