+++ source-content/hipaa/README.md
@@ -2,7 +2,7 @@
 
 ## Why this directory is empty
 
-The HIPAA Security Rule is published by the U.S. Department of Health and Human Services (HHS) and is **freely available**. However, an automated downloader for this framework has not yet been implemented in comply-with-me.
+The HIPAA Security Rule is published by the U.S. Department of Health and Human Services (HHS) and is **freely available**. However, an automated downloader for this framework has not yet been implemented in CompliGator.
 
 ## How to obtain
 
@@ -10,7 +10,7 @@ Download directly from HHS:
 
 - [HIPAA Security Rule (45 CFR Parts 160 and 164)](https://www.hhs.gov/hipaa/for-professionals/security/index.html)
 - [HIPAA Security Rule Summary](https://www.hhs.gov/hipaa/for-professionals/security/laws-regulations/index.html)
-- NIST SP 800-66r2 (HIPAA implementation guidance): download via comply-with-me's NIST downloader
+- NIST SP 800-66r2 (HIPAA implementation guidance): download via CompliGator's NIST downloader
 
 ## What to place here
 
@@ -21,4 +21,4 @@ Drop the PDF files into this directory. Suggested filenames:
 
 Once placed, these files will be picked up by the normalize command (when available).
 
-> **Note:** A future version of comply-with-me may automate this download.
+> **Note:** A future version of CompliGator may automate this download.
